# packages/contrato/openapi.yaml
openapi: 3.0.3
info:
  title: Tramitador API
  description: >-
    Bandeja de back-office para la gestion de solicitudes de clientes.
    Todas las rutas cuelgan del servidor `/api`.
  version: 0.1.0
servers:
  - url: /api
    description: Raiz de la API del tramitador.
paths:
  /salud:
    get:
      operationId: obtenerSalud
      summary: Comprueba que la API responde.
      description: >-
        Sonda de vida sin efectos secundarios. Devuelve `200` mientras el
        servicio este en pie; si no responde, el cliente debe asumir que la API
        no esta disponible.
      tags:
        - Salud
      responses:
        '200':
          description: La API esta operativa.
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
                    description: Siempre `true` cuando la API responde.
                    example: true
              example:
                ok: true
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /tipos:
    get:
      operationId: listarTipos
      summary: Lista el catalogo de tipos de solicitud.
      description: >-
        Devuelve todas las entradas del catalogo externo de tipos. La lista
        puede venir vacia; eso no es un error.
      tags:
        - Catalogos
      responses:
        '200':
          description: Catalogo de tipos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tipo'
              example:
                - id: alta-cuenta
                  nombre: Alta de cuenta
                  descripcion: Apertura de una cuenta nueva para un cliente.
                - id: cambio-titular
                  nombre: Cambio de titular
                  descripcion: Traspaso de la titularidad de una cuenta existente.
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /operadores:
    get:
      operationId: listarOperadores
      summary: Lista los operadores de back-office.
      description: >-
        Devuelve todos los operadores que pueden gestionar solicitudes. La
        lista puede venir vacia; eso no es un error.
      tags:
        - Catalogos
      responses:
        '200':
          description: Lista de operadores.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operador'
              example:
                - id: op-0042
                  nombre: Javier Nunez Prieto
                - id: op-0043
                  nombre: Marta Ruiz Delgado
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /solicitudes:
    get:
      operationId: listarSolicitudes
      summary: Lista las solicitudes, opcionalmente filtradas por estado.
      description: >-
        Devuelve todas las solicitudes. Si se indica `estado`, devuelve
        unicamente las que se encuentran en ese estado. Un filtro valido que no
        casa con ninguna solicitud produce `200` con un array vacio, no `404`.
      tags:
        - Solicitudes
      parameters:
        - $ref: '#/components/parameters/FiltroEstado'
      responses:
        '200':
          description: Solicitudes encontradas, ya filtradas si se indico `estado`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Solicitud'
              example:
                - id: 3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f
                  referencia: SOL-2026-0007
                  tipo: alta-cuenta
                  solicitante: Marta Ruiz Delgado
                  estado: enviada
                  operador: op-0042
                  creadaEn: '2026-02-11T09:14:32Z'
                  actualizadaEn: '2026-02-12T16:48:05Z'
        '400':
          description: El valor de `estado` no pertenece al enum `Estado`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                mensaje: El filtro de estado no es valido.
                detalle: >-
                  Valor recibido: 'pendiente'. Valores admitidos: borrador,
                  enviada, aprobada, rechazada.
        '500':
          $ref: '#/components/responses/ErrorInterno'

    post:
      operationId: crearSolicitud
      summary: Crea una solicitud nueva.
      description: >-
        Crea una solicitud a partir de un tipo del catalogo y un solicitante.
        La solicitud nace en estado `borrador` y sin operador asignado; el
        servidor le adjudica `id`, `referencia`, `creadaEn` y `actualizadaEn`.
      tags:
        - Solicitudes
      requestBody:
        required: true
        description: Datos de la solicitud a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuevaSolicitud'
            example:
              tipo: alta-cuenta
              solicitante: Marta Ruiz Delgado
      responses:
        '201':
          description: Solicitud creada.
          headers:
            Location:
              description: Ruta de la solicitud recien creada.
              schema:
                type: string
                format: uri-reference
                example: /api/solicitudes/3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
              example:
                id: 3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f
                referencia: SOL-2026-0007
                tipo: alta-cuenta
                solicitante: Marta Ruiz Delgado
                estado: borrador
                operador: null
                creadaEn: '2026-02-11T09:14:32Z'
                actualizadaEn: '2026-02-11T09:14:32Z'
        '400':
          description: >-
            El cuerpo no es valido: falta un campo obligatorio, el `tipo` no
            existe en el catalogo o el `solicitante` no es utilizable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                tipoDesconocido:
                  summary: El tipo no existe en el catalogo.
                  value:
                    mensaje: El tipo de solicitud no es valido.
                    detalle: >-
                      No existe ningun tipo con id 'alta-cuentas' en el catalogo.
                solicitanteNoValido:
                  summary: El solicitante no es utilizable.
                  value:
                    mensaje: El solicitante no es valido.
                    detalle: "El campo 'solicitante' no puede estar vacio."
                campoAusente:
                  summary: Falta un campo obligatorio.
                  value:
                    mensaje: El cuerpo de la peticion no es valido.
                    detalle: "Falta el campo obligatorio 'tipo'."
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /solicitudes/{id}:
    parameters:
      - $ref: '#/components/parameters/IdSolicitud'
    get:
      operationId: obtenerSolicitud
      summary: Obtiene una solicitud por su identificador.
      description: Devuelve la solicitud cuyo `id` coincide con el de la ruta.
      tags:
        - Solicitudes
      responses:
        '200':
          description: Solicitud encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
              example:
                id: 3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f
                referencia: SOL-2026-0007
                tipo: alta-cuenta
                solicitante: Marta Ruiz Delgado
                estado: enviada
                operador: op-0042
                creadaEn: '2026-02-11T09:14:32Z'
                actualizadaEn: '2026-02-12T16:48:05Z'
        '400':
          $ref: '#/components/responses/IdMalFormado'
        '404':
          $ref: '#/components/responses/SolicitudNoEncontrada'
        '500':
          $ref: '#/components/responses/ErrorInterno'
...
components:
  parameters:
    IdSolicitud:
      name: id
      in: path
      required: true
      description: Identificador unico e interno de la solicitud.
      schema:
        type: string
        format: uuid
      example: 3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f
    FiltroEstado:
      name: estado
      in: query
      required: false
      description: >-
        Si se indica, devuelve unicamente las solicitudes que estan en ese
        estado. Si se omite, no se filtra.
      schema:
        $ref: '#/components/schemas/Estado'
      example: enviada
  responses:
    SolicitudNoEncontrada:
      description: No existe ninguna solicitud con ese identificador.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            mensaje: La solicitud no existe.
            detalle: >-
              No hay ninguna solicitud con id
              '3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f'.
    IdMalFormado:
      description: El identificador de la ruta no tiene formato UUID.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            mensaje: El identificador de la solicitud no es valido.
            detalle: "Valor recibido: 'SOL-2026-0007'. Se esperaba un UUID."
    ErrorInterno:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            mensaje: Se ha producido un error inesperado.
  schemas:
    Estado:
      type: string
      description: >-
        Estado de una solicitud dentro de la maquina de estados.
        `aprobada` y `rechazada` son estados finales.
      enum:
        - borrador
        - enviada
        - aprobada
        - rechazada
      example: enviada
...
    Solicitud:
      type: object
      description: Solicitud de cliente gestionada por un operador de back-office.
      required:
        - id
        - referencia
        - tipo
        - solicitante
        - estado
        - operador
        - creadaEn
        - actualizadaEn
      properties:
        id:
          type: string
          format: uuid
          description: Identificador unico e interno de la solicitud.
          example: 3f2a9c14-8b7d-4e5a-9f01-6c2d5e8a4b3f
        referencia:
          type: string
          description: >-
            Codigo legible por humanos de la solicitud, unico y estable.
            Formato `SOL-<anyo>-<secuencial de 4 digitos>`.
          pattern: '^SOL-\d{4}-\d{4}$'
          example: SOL-2026-0007
        tipo:
          type: string
          description: >-
            Identificador del tipo de solicitud en el catalogo externo de tipos.
            Se corresponde con `Tipo.id`.
          example: alta-cuenta
        solicitante:
          type: string
          description: Cliente que presenta la solicitud.
          example: Marta Ruiz Delgado
        estado:
          allOf:
            - $ref: '#/components/schemas/Estado'
          description: Estado actual de la solicitud.
        operador:
          type: string
          nullable: true
          description: >-
            Identificador del operador de back-office asignado a la solicitud,
            correspondiente a `Operador.id`. Es `null` cuando no hay ninguno asignado.
          example: op-0042
        creadaEn:
          type: string
          format: date-time
          description: Instante de creacion de la solicitud (UTC, ISO 8601).
          example: '2026-02-11T09:14:32Z'
        actualizadaEn:
          type: string
          format: date-time
          description: >-
            Instante de la ultima modificacion de la solicitud (UTC, ISO 8601).
            Coincide con `creadaEn` mientras no se haya modificado.
          example: '2026-02-12T16:48:05Z'
    NuevaSolicitud:
      type: object
      description: Datos necesarios para crear una solicitud nueva.
      required:
        - tipo
        - solicitante
      properties:
        tipo:
          type: string
          description: >-
            Identificador del tipo de solicitud en el catalogo externo de tipos.
            Se corresponde con `Tipo.id`.
          example: alta-cuenta
        solicitante:
          type: string
          description: Cliente que presenta la solicitud.
          example: Marta Ruiz Delgado
...
    Tipo:
      type: object
      description: Entrada del catalogo externo de tipos de solicitud.
      required:
        - id
        - nombre
        - descripcion
      properties:
        id:
          type: string
          description: Identificador unico del tipo en el catalogo.
          example: alta-cuenta
        nombre:
          type: string
          description: Nombre corto del tipo, para mostrar en la interfaz.
          example: Alta de cuenta
        descripcion:
          type: string
          description: Explicacion del tipo de solicitud.
          example: Apertura de una cuenta nueva para un cliente.
    Operador:
      type: object
      description: Operador de back-office que puede gestionar solicitudes.
      required:
        - id
        - nombre
      properties:
        id:
          type: string
          description: Identificador unico del operador.
          example: op-0042
        nombre:
          type: string
          description: Nombre del operador, para mostrar en la interfaz.
          example: Javier Nunez Prieto
    Error:
      type: object
      description: Respuesta devuelta cuando una operacion no se puede completar.
      required:
        - mensaje
      properties:
        mensaje:
          type: string
          description: Descripcion del error, legible por humanos.
          example: La solicitud ya esta en un estado final.
        detalle:
          type: string
          description: Informacion adicional sobre el error, cuando existe.
          example: 'Estado actual: aprobada; accion solicitada: rechazar.'